fix types

+++ arch/ppc32/include/types.h
@@ -38,7 +38,7 @@ typedef signed long long __s64;
 
 typedef unsigned char __u8;
 typedef unsigned short __u16;
-typedef unsigned long __u32;
+typedef unsigned int __u32;
 typedef unsigned long long __u64;
 
 typedef __u32 __address;
